SingletonTheory.Services.AuthServices/Repositories/ContactDetails/OccupationNameRepository.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Web;
using ServiceStack.DataAccess;
using SingletonTheory.OrmLite;
using SingletonTheory.OrmLite.Interfaces;
using SingletonTheory.Services.AuthServices.Entities.ContactDetails;

namespace SingletonTheory.Services.AuthServices.Repositories.ContactDetails
{
	public class OccupationNameRepository
	{
		#region Fields & Properties

		private string _connectionStringName;

		#endregion Fields & Properties

		#region Constructors

		public OccupationNameRepository(string connectionStringName)
		{
			if (string.IsNullOrEmpty(connectionStringName))
				throw new ArgumentNullException("connectionStringName");

			_connectionStringName = connectionStringName;

			using (IDatabaseProvider provider = ProviderFactory.GetProvider(connectionStringName))
			{
				if (!provider.CollectionExists(typeof(OccupationNameEntity)))
					provider.CreateCollection(typeof(OccupationNameEntity));
			}
		}

		#endregion Constructors

		#region Public Methods

		public OccupationNameEntity Create(OccupationNameEntity entity)
		{
			using (IDatabaseProvider provider = ProviderFactory.GetProvider(_connectionStringName))
			{
				return provider.Insert<OccupationNameEntity>(entity);
			}
		}

		public List<OccupationNameEntity> Create(List<OccupationNameEntity> entities)
		{
			for (int i = 0; i < entities.Count; i++)
			{
				entities[i] = Create(entities[i]);
			}

			return entities;
		}

		public OccupationNameEntity Read(long id)
		{
			using (IDatabaseProvider provider = ProviderFactory.GetProvider(_connectionStringName))
			{
				return provider.SelectById<OccupationNameEntity>(id);
			}
		}

		public List<OccupationNameEntity> Read()
		{
			using (IDatabaseProvider provider = ProviderFactory.GetProvider(_connectionStringName))
			{
				return provider.Select<OccupationNameEntity>();
			}
		}

		public OccupationNameEntity Update(OccupationNameEntity entity)
		{
			using (IDatabaseProvider provider = ProviderFactory.GetProvider(_connectionStringName))
			{
				OccupationNameEntity entityToUpdate = Read(entity.Id);
				if (entityToUpdate == null)
					throw new DataAccessException("Item not found"); //  This should not happen seeing that validation should check.

				entityToUpdate = UpdateProperties(entity, entityToUpdate);

				provider.Update<OccupationNameEntity>(entityToUpdate);

				return entityToUpdate;
			}
		}

		public List<OccupationNameEntity> Update(List<OccupationNameEntity> entities)
		{
			for (int i = 0; i < entities.Count; i++)
			{
				entities[i] = Update(entities[i]);
			}

			return entities;
		}

		public OccupationNameEntity Delete(OccupationNameEntity entity)
		{
			entity.DeletedDate = DateTime.UtcNow;

			return Update(entity);
		}

		public void ClearCollection()
		{
			using (IDatabaseProvider provider = ProviderFactory.GetProvider(_connectionStringName))
			{
				if (provider.CollectionExists(typeof(OccupationNameEntity)))
					provider.DeleteAll<OccupationNameEntity>();
			}
		}

		#endregion Public Methods

		#region Private Methods

		private OccupationNameEntity UpdateProperties(OccupationNameEntity entity, OccupationNameEntity entityToUpdate)
		{
			entityToUpdate.Description = entity.Description;
			entityToUpdate.DeletedDate = entity.DeletedDate;

			return entityToUpdate;
		}

		#endregion Private Methods
	}
}